Move CFoulState traits
