+++ include/football/CFoulState.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CFoulState.h"
 
 #include "types/CPerson.h"
 #include "types/CFoulDrawConfiguration.h"
@@ -9,25 +9,7 @@
 #include "football/CMatchConfiguration.h"
 #include "football/CTeamStrategy.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CFoulState;
-}
-
-template <> struct json_traits<football::CFoulState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Foul state";
-	//! JSON key for the \copybrief football::CFoulState::mCommitter
-	static inline constexpr std::string_view COMMIITER_KEY = "Committer";
-	//! JSON key for the \copybrief football::CFoulState::mOutcome
-	static inline constexpr std::string_view OUTCOME_KEY = "Outcome";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -83,6 +65,4 @@ CFoulState::CFoulState(
 {
 }
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
